Split editing tests into different test files
