+++ core/ecschema-editing/src/test/Editing/RelationshipClasses.test.ts
@@ -1,0 +1,78 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
+import { expect } from "chai";
+import { 
+    RelationshipClass, RelationshipClassProps, RelationshipConstraintProps, Schema, SchemaContext, SchemaKey,
+} from "@bentley/ecschema-metadata";
+import { SchemaContextEditor } from "../../Editing/Editor";
+
+describe("Relationship tests from an existing schema", () => {
+    const schemaJson = {
+      $schema: "https://dev.bentley.com/json_schemas/ec/32/ecschema",
+      name: "TestSchema",
+      version: "1.2.3",
+      items: {
+        SourceBaseEntity: {
+          schemaItemType: "EntityClass",
+        },
+        TargetBaseEntity: {
+          schemaItemType: "EntityClass",
+        },
+        TestSourceEntity: {
+          schemaItemType: "EntityClass",
+          baseClass: "TestSchema.SourceBaseEntity",
+        },
+        TestTargetEntity: {
+          schemaItemType: "EntityClass",
+          baseClass: "TestSchema.TargetBaseEntity",
+        },
+      },
+    };
+    let testEditor: SchemaContextEditor;
+    let testSchema: Schema;
+    let testKey: SchemaKey;
+    let context: SchemaContext;
+
+    beforeEach(async () => {
+      context = new SchemaContext();
+      testSchema = await Schema.fromJson(schemaJson, context);
+      testEditor = new SchemaContextEditor(context);
+      testKey = testSchema.schemaKey;
+    });
+
+    it("should create a relationship class given a valid RelationshipClassProps", async () => {
+      const sourceJson: RelationshipConstraintProps = {
+        polymorphic: true,
+        multiplicity: "(0..*)",
+        roleLabel: "Source RoleLabel",
+        abstractConstraint: "TestSchema.SourceBaseEntity",
+        constraintClasses: [
+          "TestSchema.TestSourceEntity",
+        ],
+      };
+      const targetJson: RelationshipConstraintProps = {
+        polymorphic: true,
+        multiplicity: "(0..*)",
+        roleLabel: "Target RoleLabel",
+        abstractConstraint: "TestSchema.TargetBaseEntity",
+        constraintClasses: [
+          "TestSchema.TestTargetEntity",
+        ],
+      };
+
+      const relClassProps: RelationshipClassProps = {
+        name: "TestRelationship",
+        strength: "Embedding",
+        strengthDirection: "Forward",
+        source: sourceJson,
+        target: targetJson,
+      };
+
+      const result = await testEditor.relationships.createFromProps(testKey, relClassProps);
+      const relClass = await testEditor.schemaContext.getSchemaItem(result.itemKey!) as RelationshipClass;
+      const baseSourceClassKey = testSchema.getSchemaItemKey("TestSchema.SourceBaseEntity");
+      expect(await relClass.source.abstractConstraint).to.eql(await testEditor.schemaContext.getSchemaItem(baseSourceClassKey));
+    });
+  });
